update example command
