+++ coordinates_conversion/bin/fasta_diff.py
@@ -29,7 +29,7 @@ def fasta_file_to_dict(fasta_file, id=True, header=False, seq=False):
     Duplicate keys are checked and a warning is logged if found.
     The value of fasta_dict is a python dict with 3 keys: header, id and seq
     """
-   
+
     fasta_file_f = fasta_file
     if isinstance(fasta_file, str):
         fasta_file_f = open(fasta_file, 'rb')
@@ -185,7 +185,7 @@ def fasta_diff(old_fasta_file, new_fasta_file, debug=True, header_check=False, r
         old_seqs = old_fasta_dict.keys()
         new_seqs = new_fasta_dict.keys()
         match_split = dict() # {matches[0]: {'matches': {new_seq}, 'alignment': [oldid, oldstart, oldend, newid, newstart, newend]}
-        
+
         match_split_order = list()
         for new_seq in new_seqs:
             segments = new_seq.replace('N', ' ').split()
@@ -245,7 +245,6 @@ def fasta_diff(old_fasta_file, new_fasta_file, debug=True, header_check=False, r
                         if new_seq[tmp_newend] == nucl:
                             tmp_oldend += 1
                             tmp_newend += 1
-          
                         else:
                             break
                     except:
@@ -263,23 +262,22 @@ def fasta_diff(old_fasta_file, new_fasta_file, debug=True, header_check=False, r
 
             elif len(matches) > 1:
                 logging.warning('Failed one to one mapping: %s has %d matches: %s\n' % (new_fasta_dict[new_seq]['id'], len(matches), ','.join([x[0][0][0].split('.')[0] for x in matches])))
-        
+
         onetomultiple.update(match_split)
         for match in match_split_order:
             if len(match_split[match]['matches']) == 1:
                 # one to one
-                alignment_list.extend(match_split[match]['alignment'])        
+                alignment_list.extend(match_split[match]['alignment'])
                 del old_fasta_dict[match] # matches[0]
                 del new_fasta_dict[match_split[match]['matches'][0]] # new_seq
                 del match_split[match]
 
     def one_to_multiple_match():
-        
         stagelist=list()
         for match in onetomultiple:
             # one to mutiple
             if len(onetomultiple[match]['matches']) > 1:
-                stagelist.extend(onetomultiple[match]['alignment'])    
+                stagelist.extend(onetomultiple[match]['alignment'])
 
                 overlap=dict()
                 #Fetch old_id and new_id
@@ -288,7 +286,6 @@ def fasta_diff(old_fasta_file, new_fasta_file, debug=True, header_check=False, r
                     if pair not in overlap:
                        overlap[pair]=set()
                     overlap[pair].update((tmp[1],tmp[2]))
-                
                 pairs=overlap.keys()
                 pairs_sort=sorted(pairs, key=lambda aaa:aaa[0])
                 run_sort=set()
@@ -300,24 +297,24 @@ def fasta_diff(old_fasta_file, new_fasta_file, debug=True, header_check=False, r
                         if pair1==pair2:
                             continue
                         else:
-                            if (pair1,pair2) in run_sort: continue
+                            if (pair1,pair2) in run_sort:
+                                continue
                             run_sort.update([(pair1,pair2),(pair2,pair1)])
-                            
                             if pair1[0]==pair2[0]:
                                 # find overlap pair
                                 if (min(overlap[pair1])<=min(overlap[pair2]) and min(overlap[pair2])<=max(overlap[pair1]))\
                                 or (min(overlap[pair1])<=max(overlap[pair2]) and max(overlap[pair2])<=max(overlap[pair1])):
-                                
+
                                     delete_pairs.update((pair1,pair2))
-                                    # add overlap pair to delete_pairs      
-                                             
-                stage_four_result=list() 
+                                    # add overlap pair to delete_pairs
+
+                stage_four_result=list()
                 for delete in stagelist:
                    if (delete[0],delete[3]) not in delete_pairs:
                        stage_four_result.append(delete)
 
                        if match in old_fasta_dict:
-                            del old_fasta_dict[match] 
+                            del old_fasta_dict[match]
 
                        for new in onetomultiple[match]['matches']:
                            if new in new_fasta_dict:
@@ -327,7 +324,7 @@ def fasta_diff(old_fasta_file, new_fasta_file, debug=True, header_check=False, r
         if onetomultiple:
             alignment_list.extend(stage_four_result)
         # add empty to final result
-                                                 
+
     stages = [match_identical_sequence, match_truncated_sequence, match_split_subsequence,one_to_multiple_match]
     matched_sequence_count = 0
     for stage in range(len(stages)):
@@ -369,8 +366,8 @@ def fasta_diff(old_fasta_file, new_fasta_file, debug=True, header_check=False, r
         logging.info('  Matched sequences: %d (New : %d)', new_matched_sequence_count, new_matched_sequence_count - matched_sequence_count)
         logging.info('  Unmatched sequences in old FASTA: %d', len(old_fasta_dict))
         logging.info('  Unmatched sequences in new FASTA: %d', len(new_fasta_dict))
-        
-        
+
+
         matched_sequence_count = new_matched_sequence_count
         if debug:
             fasta_dict_to_file(old_fasta_dict, old_fasta_file + '_stage_' + str(stage + 1) + '_unmatched')
@@ -395,6 +392,9 @@ def main():
     Stage 3: Find cases where part of the sequence was converted into Ns
     Outputs the 6 columns as tab-separated values: old_id, old_start, old_end, new_id, new_start, new_end
     Originally created to compare the NCBI version to the original reference of Diaphorina citri
+
+    Example:
+        fasta_diff example_file/old.fa example_file/new.fa -o match.tsv -r report.txt
     """))
     parser.add_argument('old_fasta', type=str, help='The original FASTA file')
     parser.add_argument('new_fasta', type=str, help='The new FASTA file')
