+++ coordinates_conversion/bin/update_bed.py
@@ -147,9 +147,9 @@ def main():
     are removed or replaced with Ns, are written to a new file with '_removed'(default) appended to the original Bed file name.
 
     Example:
-        fasta_diff old.fa new.fa | %(prog)s a.Bed b.Bed c.Bed
-    """))             
-    parser.add_argument('Bed_files', metavar='Bed_FILE', nargs='+', type=str, 
+        fasta_diff example_file/old.fa example_file/new.fa | %(prog)s example_file/example.bed
+    """))
+    parser.add_argument('Bed_files', metavar='Bed_FILE', nargs='+', type=str,
     help='List one or more Bed files to be updated')
     parser.add_argument('-a', '--alignment_file', type=argparse.FileType('rb'), default=sys.stdin,help='The alignment file generated by fasta_diff, a TSV file with 6 columns: old_id, old_start, old_end, new_id, new_start, new_end (default: STDIN)')
