+++ coordinates_conversion/bin/update_bedgraph.py
@@ -113,9 +113,9 @@ def main():
     are removed or replaced with Ns, are written to a new file with '_removed'(default) appended to the original BedGraph file name.
 
     Example:
-        fasta_diff old.fa new.fa | %(prog)s a.BedGraph b.BedGraph c.BedGraph
-    """))             
-    parser.add_argument('BedGraph_files', metavar='BedGraph_FILE', nargs='+', type=str, 
+        fasta_diff example_file/old.fa example_file/new.fa | %(prog)s example_file/example.bedGraph
+    """))
+    parser.add_argument('BedGraph_files', metavar='BedGraph_FILE', nargs='+', type=str,
     help='List one or more BedGraph files to be updated')
     parser.add_argument('-a', '--alignment_file', type=argparse.FileType('rb'), default=sys.stdin,help='The alignment file generated by fasta_diff, a TSV file with 6 columns: old_id, old_start, old_end, new_id, new_start, new_end (default: STDIN)')
